+++ lib/ArangoDBClient/CollectionHandler.php
@@ -58,6 +58,11 @@ class CollectionHandler extends Handler
     const OPTION_KEYS = 'keys';
 
     /**
+     * stream parameter
+     */
+    const OPTION_STREAM = 'stream';
+
+    /**
      * left parameter
      */
     const OPTION_LEFT = 'left';
@@ -201,6 +206,16 @@ class CollectionHandler extends Handler
      * revision option
      */
     const OPTION_REVISION = 'revision';
+    
+    /**
+     * responsible shard option
+     */
+    const OPTION_RESPONSIBLE_SHARD = 'responsibleShard';
+    
+    /**
+     * shards option
+     */
+    const OPTION_SHARDS = 'shards';
 
     /**
      * properties option
@@ -485,7 +500,7 @@ class CollectionHandler extends Handler
      *
      * @throws Exception
      *
-     * @param mixed   $collectionId         - collection id as a string or number
+     * @param mixed   $collection           - collection as string or object
      * @param boolean $withRevisions        - optional boolean whether or not to include document revision ids
      *                                      in the checksum calculation.
      * @param boolean $withData             - optional boolean whether or not to include document body data in the
@@ -493,10 +508,9 @@ class CollectionHandler extends Handler
      *
      * @return array - array containing keys "checksum" and "revision"
      */
-    public function getChecksum($collectionId, $withRevisions = false, $withData = false)
+    public function getChecksum($collection, $withRevisions = false, $withData = false)
     {
-
-        $url      = UrlHelper::buildUrl(Urls::URL_COLLECTION, [$collectionId, self::OPTION_CHECKSUM]);
+        $url      = UrlHelper::buildUrl(Urls::URL_COLLECTION, [$this->makeCollection($collection), self::OPTION_CHECKSUM]);
         $url      = UrlHelper::appendParamsUrl($url, ['withRevisions' => $withRevisions, 'withData' => $withData]);
         $response = $this->getConnection()->get($url);
 
@@ -511,14 +525,13 @@ class CollectionHandler extends Handler
      *
      * @throws Exception
      *
-     * @param mixed $collectionId - collection id as a string or number
+     * @param mixed $collection - collection as string or object
      *
      * @return array - containing a key revision
      */
-    public function getRevision($collectionId)
+    public function getRevision($collection)
     {
-
-        $url      = UrlHelper::buildUrl(Urls::URL_COLLECTION, [$collectionId, self::OPTION_REVISION]);
+        $url      = UrlHelper::buildUrl(Urls::URL_COLLECTION, [$this->makeCollection($collection), self::OPTION_REVISION]);
         $response = $this->getConnection()->get($url);
 
         return $response->getJson();
@@ -529,22 +542,16 @@ class CollectionHandler extends Handler
      *
      * @throws Exception
      *
-     * @param mixed  $collection - collection id as string or number or collection object
+     * @param mixed  $collection - collection as string or object
      * @param string $name       - new name for collection
      *
      * @return bool - always true, will throw if there is an error
      */
     public function rename($collection, $name)
     {
-        $collectionId = $this->getCollectionId($collection);
-
-        if ($this->isValidCollectionId($collectionId)) {
-            throw new ClientException('Cannot alter a collection without a collection id');
-        }
-
         $params = [Collection::ENTRY_NAME => $name];
         $this->getConnection()->put(
-            UrlHelper::buildUrl(Urls::URL_COLLECTION, [$collectionId, self::OPTION_RENAME]),
+            UrlHelper::buildUrl(Urls::URL_COLLECTION, [$this->makeCollection($collection), self::OPTION_RENAME]),
             $this->json_encode_wrapper($params)
         );
 
@@ -558,20 +565,14 @@ class CollectionHandler extends Handler
      *
      * @throws Exception
      *
-     * @param mixed $collection - collection id as string or number or collection object
+     * @param mixed $collection - collection as string or object
      *
      * @return HttpResponse - HTTP response object
      */
     public function load($collection)
     {
-        $collectionId = $this->getCollectionId($collection);
-
-        if ($this->isValidCollectionId($collectionId)) {
-            throw new ClientException('Cannot alter a collection without a collection id');
-        }
-
         $result = $this->getConnection()->put(
-            UrlHelper::buildUrl(Urls::URL_COLLECTION, [$collectionId, self::OPTION_LOAD]),
+            UrlHelper::buildUrl(Urls::URL_COLLECTION, [$this->makeCollection($collection), self::OPTION_LOAD]),
             ''
         );
 
@@ -585,20 +586,14 @@ class CollectionHandler extends Handler
      *
      * @throws Exception
      *
-     * @param mixed $collection - collection id as string or number or collection object
+     * @param mixed $collection - collection as string or object
      *
      * @return HttpResponse - HTTP response object
      */
     public function unload($collection)
     {
-        $collectionId = $this->getCollectionId($collection);
-
-        if ($this->isValidCollectionId($collectionId)) {
-            throw new ClientException('Cannot alter a collection without a collection id');
-        }
-
         $result = $this->getConnection()->put(
-            UrlHelper::buildUrl(Urls::URL_COLLECTION, [$collectionId, self::OPTION_UNLOAD]),
+            UrlHelper::buildUrl(Urls::URL_COLLECTION, [$this->makeCollection($collection), self::OPTION_UNLOAD]),
             ''
         );
 
@@ -650,18 +645,12 @@ class CollectionHandler extends Handler
      */
     public function drop($collection, array $options = [])
     {
-        $collectionName = $this->getCollectionName($collection);
-
-        if ($this->isValidCollectionId($collectionName)) {
-            throw new ClientException('Cannot alter a collection without a collection id');
-        }
-
         $appendix = '';
         if (is_array($options) && isset($options['isSystem'])) {
             $appendix = '?isSystem=' . UrlHelper::getBoolString($options['isSystem']);
         }
 
-        $this->getConnection()->delete(UrlHelper::buildUrl(Urls::URL_COLLECTION, [$collectionName]) . $appendix);
+        $this->getConnection()->delete(UrlHelper::buildUrl(Urls::URL_COLLECTION, [$this->makeCollection($collection)]) . $appendix);
 
         return true;
     }
@@ -766,7 +755,7 @@ class CollectionHandler extends Handler
      *
      * @throws Exception
      *
-     * @param mixed $collectionId   - collection id as string or number
+     * @param mixed $collectionId   - collection as string or object
      * @param mixed $importFileName - The filename that holds the import data.
      * @param array $options        - optional - an array of options.
      *                              <p>Options are :<br>
@@ -784,11 +773,7 @@ class CollectionHandler extends Handler
      *
      * @return array - returns an array with the server's response data from the import command
      */
-    public function importFromFile(
-        $collectionId,
-        $importFileName,
-        array $options = []
-    )
+    public function importFromFile($collection, $importFileName, array $options = [])
     {
 
         $contents = file_get_contents($importFileName);
@@ -796,7 +781,7 @@ class CollectionHandler extends Handler
             throw new ClientException('Input file "' . $importFileName . '" could not be found.');
         }
 
-        return $this->import($collectionId, $contents, $options);
+        return $this->import($collection, $contents, $options);
     }
 
 
@@ -806,7 +791,7 @@ class CollectionHandler extends Handler
      * This will throw on all errors except insertion errors
      *
      *
-     * @param              $collection   mixed $collection - collection id as string or number
+     * @param              $collection   mixed $collection - collection as string or object
      * @param string|array $importData   - The data to import. This can be a string holding the data according to the type of import, or an array of documents
      * @param array        $options      - optional - an array of options.
      *                                   <p>Options are :<br>
@@ -832,11 +817,7 @@ class CollectionHandler extends Handler
      * @throws \ArangoDBClient\Exception
      * @throws \ArangoDBClient\ClientException
      */
-    public function import(
-        $collection,
-        $importData,
-        array $options = []
-    )
+    public function import($collection, $importData, array $options = [])
     {
         $collection = $this->makeCollection($collection);
 
@@ -880,20 +861,23 @@ class CollectionHandler extends Handler
     /**
      * Create a hash index
      *
-     * @param string  $collectionId - the collection id
+     * @param mixed   $collection   - the collection as name or object
      * @param array   $fields       - an array of fields
      * @param bool    $unique       - whether the values in the index should be unique or not
      * @param bool    $sparse       - whether the index should be sparse
      * @param bool    $inBackground - true if index shall be created in background
      *
-     * @link https://www.arangodb.com/docs/devel/indexing-hash.html
+     * @deprecated use CollectionHandler::createIndex instead
      *
      * @return array - server response of the created index
      * @throws \ArangoDBClient\Exception
      */
-    public function createHashIndex($collectionId, array $fields, $unique = null, $sparse = null, $inBackground = false)
+    public function createHashIndex($collection, array $fields, $unique = null, $sparse = null, $inBackground = false)
     {
-        $indexOptions = [];
+        $indexOptions = [
+            self::OPTION_TYPE   => 'hash',
+            self::OPTION_FIELDS => $fields
+        ];
 
         if ($unique) {
             $indexOptions[self::OPTION_UNIQUE] = (bool) $unique;
@@ -905,25 +889,28 @@ class CollectionHandler extends Handler
             $indexOptions[self::OPTION_IN_BACKGROUND] = (bool) $inBackground;
         }
 
-        return $this->index($collectionId, self::OPTION_HASH_INDEX, $fields, null, $indexOptions);
+        return $this->createIndex($collection, $indexOptions);
     }
 
     /**
      * Create a fulltext index
      *
-     * @param string $collectionId - the collection id
+     * @param mixed  $collection   - collection as string or object
      * @param array  $fields       - an array of fields
      * @param int    $minLength    - the minimum length of words to index
      * @param bool   $inBackground - true if index shall be created in background
      *
-     * @link https://www.arangodb.com/docs/devel/indexing-fulltext.html
+     * @deprecated use CollectionHandler::createIndex instead
      *
      * @return array - server response of the created index
      * @throws \ArangoDBClient\Exception
      */
-    public function createFulltextIndex($collectionId, array $fields, $minLength = null, $inBackground = false)
+    public function createFulltextIndex($collection, array $fields, $minLength = null, $inBackground = false)
     {
-        $indexOptions = [];
+        $indexOptions = [
+            self::OPTION_TYPE   => 'fulltext',
+            self::OPTION_FIELDS => $fields
+        ];
 
         if ($minLength) {
             $indexOptions[self::OPTION_MIN_LENGTH] = $minLength;
@@ -932,26 +919,29 @@ class CollectionHandler extends Handler
             $indexOptions[self::OPTION_IN_BACKGROUND] = (bool) $inBackground;
         }
 
-        return $this->index($collectionId, self::OPTION_FULLTEXT_INDEX, $fields, null, $indexOptions);
+        return $this->createIndex($collection, $indexOptions);
     }
 
     /**
      * Create a skip-list index
      *
-     * @param string $collectionId - the collection id
+     * @param mixed  $collection   - collection as string or object
      * @param array  $fields       - an array of fields
      * @param bool   $unique       - whether the index is unique or not
      * @param bool   $sparse       - whether the index should be sparse
      * @param bool   $inBackground - true if index shall be created in background
      *
-     * @link https://www.arangodb.com/docs/devel/indexing-skiplist.html
+     * @deprecated use CollectionHandler::createIndex instead
      *
      * @return array - server response of the created index
      * @throws \ArangoDBClient\Exception
      */
-    public function createSkipListIndex($collectionId, array $fields, $unique = null, $sparse = null, $inBackground = false)
+    public function createSkipListIndex($collection, array $fields, $unique = null, $sparse = null, $inBackground = false)
     {
-        $indexOptions = [];
+        $indexOptions = [
+            self::OPTION_TYPE   => 'skiplist',
+            self::OPTION_FIELDS => $fields
+        ];
 
         if ($unique) {
             $indexOptions[self::OPTION_UNIQUE] = (bool) $unique;
@@ -963,26 +953,29 @@ class CollectionHandler extends Handler
             $indexOptions[self::OPTION_IN_BACKGROUND] = (bool) $inBackground;
         }
 
-        return $this->index($collectionId, self::OPTION_SKIPLIST_INDEX, $fields, null, $indexOptions);
+        return $this->createIndex($collection, $indexOptions);
     }
 
     /**
      * Create a persistent index
      *
-     * @param string $collectionId - the collection id
+     * @param mixed  $collection   - collection as string or object
      * @param array  $fields       - an array of fields
      * @param bool   $unique       - whether the index is unique or not
      * @param bool   $sparse       - whether the index should be sparse
-     * @param bool    $inBackground - true if index shall be created in background
+     * @param bool   $inBackground - true if index shall be created in background
      *
-     * @link https://www.arangodb.com/docs/devel/indexing-persistent.html
+     * @deprecated use CollectionHandler::createIndex instead
      *
      * @return array - server response of the created index
      * @throws \ArangoDBClient\Exception
      */
-    public function createPersistentIndex($collectionId, array $fields, $unique = null, $sparse = null, $inBackground = false)
+    public function createPersistentIndex($collection, array $fields, $unique = null, $sparse = null, $inBackground = false)
     {
-        $indexOptions = [];
+        $indexOptions = [
+            self::OPTION_TYPE   => 'persistent',
+            self::OPTION_FIELDS => $fields
+        ];
 
         if ($unique) {
             $indexOptions[self::OPTION_UNIQUE] = (bool) $unique;
@@ -994,50 +987,56 @@ class CollectionHandler extends Handler
             $indexOptions[self::OPTION_IN_BACKGROUND] = (bool) $inBackground;
         }
 
-        return $this->index($collectionId, self::OPTION_PERSISTENT_INDEX, $fields, null, $indexOptions);
+        return $this->createIndex($collection, $indexOptions);
     }
     
     /**
      * Create a TTL index
      *
-     * @param string $collectionId - the collection id
+     * @param mixed  $collection   - collection as string or object
      * @param array  $fields       - an array of fields (only a single one allowed)
      * @param number $expireAfter  - number of seconds after index value after which documents expire
-     * @param bool    $inBackground - true if index shall be created in background
+     * @param bool   $inBackground - true if index shall be created in background
      *
-     * @link https://www.arangodb.com/docs/devel/indexing-ttl.html
+     * @deprecated use CollectionHandler::createIndex instead
      *
      * @return array - server response of the created index
      * @throws \ArangoDBClient\Exception
      */
-    public function createTtlIndex($collectionId, array $fields, $expireAfter, $inBackground = false)
+    public function createTtlIndex($collection, array $fields, $expireAfter, $inBackground = false)
     {
         $indexOptions = [
-          self::OPTION_EXPIRE_AFTER => (double) $expireAfter
+            self::OPTION_TYPE         => 'ttl',
+            self::OPTION_FIELDS       => $fields,
+            self::OPTION_EXPIRE_AFTER => (double) $expireAfter
         ];
+    
         if ($inBackground) {
             $indexOptions[self::OPTION_IN_BACKGROUND] = (bool) $inBackground;
         }
 
-        return $this->index($collectionId, self::OPTION_TTL_INDEX, $fields, null, $indexOptions);
+        return $this->createIndex($collection, $indexOptions);
     }
 
     /**
      * Create a geo index
      *
-     * @param string  $collectionId - the collection id
+     * @param mixed   $collection   - collection as string or object
      * @param array   $fields       - an array of fields
      * @param bool    $geoJson      - whether to use geoJson or not
      * @param bool    $inBackground - true if index shall be created in background
      *
-     * @link https://www.arangodb.com/docs/devel/indexing-geo.html
+     * @deprecated use CollectionHandler::createIndex instead
      *
      * @return array - server response of the created index
      * @throws \ArangoDBClient\Exception
      */
-    public function createGeoIndex($collectionId, array $fields, $geoJson = null, $inBackground = false)
+    public function createGeoIndex($collection, array $fields, $geoJson = null, $inBackground = false)
     {
-        $indexOptions = [];
+        $indexOptions = [
+            self::OPTION_TYPE   => 'geo',
+            self::OPTION_FIELDS => $fields,
+        ];
 
         if ($geoJson) {
             $indexOptions[self::OPTION_GEOJSON] = (bool) $geoJson;
@@ -1046,7 +1045,7 @@ class CollectionHandler extends Handler
             $indexOptions[self::OPTION_IN_BACKGROUND] = (bool) $inBackground;
         }
 
-        return $this->index($collectionId, self::OPTION_GEO_INDEX, $fields, null, $indexOptions);
+        return $this->createIndex($collection, $indexOptions);
     }
 
     /**
@@ -1058,18 +1057,19 @@ class CollectionHandler extends Handler
      *
      * @throws Exception
      *
-     * @param mixed  $collectionId - The id of the collection where the index is to be created
+     * @param mixed  $collection   - collection as string or object
      * @param string $type         - index type: hash, skiplist, geo, ttl, fulltext, or persistent
      * @param array  $attributes   - an array of attributes that can be defined like array('a') or array('a', 'b.c')
      * @param bool   $unique       - true/false to create a unique index
      * @param array  $indexOptions - an associative array of options for the index like array('geoJson' => true, 'sparse' => false)
      *
+     * @deprecated use CollectionHandler::createIndex instead
+     *
      * @return array - server response of the created index
      */
-    public function index($collectionId, $type, array $attributes = [], $unique = false, array $indexOptions = [])
+    public function index($collection, $type, array $attributes = [], $unique = false, array $indexOptions = [])
     {
-
-        $urlParams  = [self::OPTION_COLLECTION => $collectionId];
+        $urlParams  = [self::OPTION_COLLECTION => $this->makeCollection($collection)];
         $bodyParams = [
             self::OPTION_TYPE   => $type,
             self::OPTION_FIELDS => $attributes,
@@ -1087,8 +1087,43 @@ class CollectionHandler extends Handler
         $httpCode = $response->getHttpCode();
         switch ($httpCode) {
             case 404:
-                throw new ClientException('Collection-identifier is unknown');
+                throw new ClientException('Collection is unknown');
+                break;
+            case 400:
+                throw new ClientException('cannot create unique index due to documents violating uniqueness');
+                break;
+        }
 
+        return $response->getJson();
+    }
+    
+    
+    /**
+     * Creates an index on a collection on the server
+     *
+     * This will create an index on the collection on the server and return its id
+     *
+     * This will throw if the index cannot be created
+     *
+     * @throws Exception
+     *
+     * @param mixed  $collection   - collection as string or object
+     * @param array  $indexOptions - an associative array of options for the index like array('type' => hash, 'fields' => ..., 'sparse' => false)
+     *
+     * @return array - server response of the created index
+     * @since 3.5
+     */
+    public function createIndex($collection, array $indexOptions)
+    {
+        $urlParams  = [self::OPTION_COLLECTION => $this->makeCollection($collection)];
+
+        $url      = UrlHelper::appendParamsUrl(Urls::URL_INDEX, $urlParams);
+        $response = $this->getConnection()->post($url, $this->json_encode_wrapper($indexOptions));
+
+        $httpCode = $response->getHttpCode();
+        switch ($httpCode) {
+            case 404:
+                throw new ClientException('Collection is unknown');
                 break;
             case 400:
                 throw new ClientException('cannot create unique index due to documents violating uniqueness');
@@ -1125,13 +1160,13 @@ class CollectionHandler extends Handler
      *
      * @throws Exception
      *
-     * @param mixed $collectionId - collection id as a string or number
+     * @param mixed $collection - collection as string or object
      *
      * @return array $data - the indexes result-set from the server
      */
-    public function getIndexes($collectionId)
+    public function getIndexes($collection)
     {
-        $urlParams = [self::OPTION_COLLECTION => $collectionId];
+        $urlParams = [self::OPTION_COLLECTION => $this->makeCollection($collection)];
         $url       = UrlHelper::appendParamsUrl(Urls::URL_INDEX, $urlParams);
         $response  = $this->getConnection()->get($url);
 
@@ -1143,16 +1178,75 @@ class CollectionHandler extends Handler
      *
      * @throws Exception
      *
+     * @param mixed $collection - collection as string or object
      * @param mixed $indexHandle - index handle (collection name / index id)
      *
      * @return bool - always true, will throw if there is an error
      */
-    public function dropIndex($indexHandle)
+    public function dropIndex($collection, $indexHandle = null)
     {
-        $handle = explode('/', $indexHandle);
-        $this->getConnection()->delete(UrlHelper::buildUrl(Urls::URL_INDEX, [$handle[0], $handle[1]]));
+        if ($indexHandle === null) {
+            $handle = explode('/', $collection);
+        } else {
+            $handle = [ $this->makeCollection($collection), $indexHandle ];
+        }
+
+        if (count($handle) > 2) {
+          throw new ClientException('Invalid index handle');
+        }
+
+        $this->getConnection()->delete(UrlHelper::buildUrl(Urls::URL_INDEX, $handle));
 
         return true;
+    }
+    
+    
+    /**
+     * Get the responsible shard for a document
+     *
+     * @throws Exception
+     *
+     * @param mixed $collection - collection as string or object
+     * @param mixed $document   - document
+     *
+     * @return string - shard id
+     * @since 3.5
+     */
+    public function getResponsibleShard($collection, $document)
+    {
+        if (is_array($document)) {
+            $data = $document;
+        } else {
+            $data = $document->getAll(['_includeInternals' => true ]);
+        }
+
+        $collection = $this->makeCollection($collection);
+        $url        = UrlHelper::buildUrl(Urls::URL_COLLECTION, [$collection, self::OPTION_RESPONSIBLE_SHARD]);
+        $response = $this->getConnection()->put($url, $this->json_encode_wrapper($data));
+        $data     = $response->getJson();
+
+        return $data['shardId'];
+    }
+    
+    
+    /**
+     * Get the shards of a collection
+     *
+     * @throws Exception
+     *
+     * @param mixed $collection - collection as string or object
+     *
+     * @return array - array with shard ids
+     * @since 3.5
+     */
+    public function getShards($collection) 
+    {
+        $collection = $this->makeCollection($collection);
+        $url        = UrlHelper::buildUrl(Urls::URL_COLLECTION, [$collection, self::OPTION_SHARDS]);
+        $response = $this->getConnection()->get($url);
+        $data     = $response->getJson();
+
+        return $data['shards'];
     }
 
     /**
@@ -1160,20 +1254,19 @@ class CollectionHandler extends Handler
      *
      * This will throw if the document cannot be fetched from the server
      *
-     *
      * @throws Exception
      *
-     * @param mixed $collectionId - collection id as string or number
+     * @param mixed $collection - collection as string or object
      *
      * @return Document - the document fetched from the server
      * @since 1.2
      */
-    public function any($collectionId)
+    public function any($collection)
     {
         $_documentClass = $this->_documentClass;
 
         $data = [
-            self::OPTION_COLLECTION => $collectionId,
+            self::OPTION_COLLECTION => $this->makeCollection($collection),
         ];
 
         $response = $this->getConnection()->put(Urls::URL_ANY, $this->json_encode_wrapper($data));
@@ -1190,7 +1283,7 @@ class CollectionHandler extends Handler
     /**
      * Returns all documents of a collection
      *
-     * @param mixed $collectionId     - collection id as string or number
+     * @param mixed $collection       - collection as string or object
      * @param array $options          - optional array of options.
      *                                <p>Options are :<br>
      *                                <li>'_sanitize'         - True to remove _id and _rev attributes from result documents. Defaults to false.</li>
@@ -1211,10 +1304,10 @@ class CollectionHandler extends Handler
      * @throws \ArangoDBClient\Exception
      * @throws \ArangoDBClient\ClientException
      */
-    public function all($collectionId, array $options = [])
+    public function all($collection, array $options = [])
     {
         $body = [
-            self::OPTION_COLLECTION => $collectionId,
+            self::OPTION_COLLECTION => $this->makeCollection($collection),
         ];
 
         $body = $this->includeOptionsInBody(
@@ -1225,6 +1318,8 @@ class CollectionHandler extends Handler
                 self::OPTION_SKIP  => null,
             ]
         );
+        
+        $body[self::OPTION_STREAM] = true;
 
         $response = $this->getConnection()->put(Urls::URL_ALL, $this->json_encode_wrapper($body));
         
@@ -1278,7 +1373,7 @@ class CollectionHandler extends Handler
      *
      * @throws Exception
      *
-     * @param mixed $collectionId      - collection id as string or number
+     * @param mixed $collection        - collection as string or object
      * @param mixed $document          - the example document as a Document object or an array
      * @param array $options           - optional, prior to v1.0.0 this was a boolean value for sanitize, since v1.0.0 it's an array of options.
      *                                 <p>Options are :<br>
@@ -1297,7 +1392,7 @@ class CollectionHandler extends Handler
      *
      * @return cursor - Returns a cursor containing the result
      */
-    public function byExample($collectionId, $document, array $options = [])
+    public function byExample($collection, $document, array $options = [])
     {
         $_documentClass = $this->_documentClass;
 
@@ -1310,7 +1405,7 @@ class CollectionHandler extends Handler
         }
 
         $body = [
-            self::OPTION_COLLECTION => $collectionId,
+            self::OPTION_COLLECTION => $this->makeCollection($collection),
             self::OPTION_EXAMPLE    => $document->getAllAsObject(['_ignoreHiddenAttributes' => true])
         ];
 
@@ -1347,7 +1442,7 @@ class CollectionHandler extends Handler
      *
      * @throws Exception
      *
-     * @param mixed $collectionId      - collection id as string or number
+     * @param mixed $collection        - collection as string or object
      * @param mixed $document          - the example document as a Document object or an array
      * @param array $options           - optional, an array of options.
      *                                 <p>Options are :<br>
@@ -1364,7 +1459,7 @@ class CollectionHandler extends Handler
      * @return Document - the document fetched from the server
      * @since 1.2
      */
-    public function firstExample($collectionId, $document, array $options = [])
+    public function firstExample($collection, $document, array $options = [])
     {
         $_documentClass = $this->_documentClass;
 
@@ -1377,7 +1472,7 @@ class CollectionHandler extends Handler
         }
 
         $data = [
-            self::OPTION_COLLECTION => $collectionId,
+            self::OPTION_COLLECTION => $this->makeCollection($collection),
             self::OPTION_EXAMPLE    => $document->getAll(['_ignoreHiddenAttributes' => true])
         ];
 
@@ -1423,12 +1518,14 @@ class CollectionHandler extends Handler
      *                                 <li>'index'     - If given, the identifier of the fulltext-index to use.</li>
      *                                 </p>
      *
+     * @deprecated use AQL queries instead
+     *
      * @return cursor - Returns a cursor containing the result
      */
     public function fulltext($collection, $attribute, $query, array $options = [])
     {
         $body = [
-            self::OPTION_COLLECTION => $collection,
+            self::OPTION_COLLECTION => $this->makeCollection($collection),
             self::OPTION_ATTRIBUTE  => $attribute,
             self::OPTION_QUERY      => $query,
         ];
@@ -1465,7 +1562,7 @@ class CollectionHandler extends Handler
      *
      * @throws Exception
      *
-     * @param mixed $collectionId - collection id as string or number
+     * @param mixed $collection   - collection as string or number
      * @param mixed $example      - the example document as a Document object or an array
      * @param mixed $newValue     - patch document or array which contains the attributes and values to be updated
      * @param mixed $options      - optional, array of options (see below) or the boolean value for $policy (for compatibility prior to version 1.1 of this method)
@@ -1475,10 +1572,12 @@ class CollectionHandler extends Handler
      *                            <li>'limit'       - can be used set a limit on how many documents to update at most. If limit is specified but is less than the number of documents in the collection, it is undefined which of the documents will be updated.</li>
      *                            </p>
      *
+     * @deprecated use AQL queries instead
+     *
      * @return bool - always true, will throw if there is an error
      * @since 1.2
      */
-    public function updateByExample($collectionId, $example, $newValue, array $options = [])
+    public function updateByExample($collection, $example, $newValue, array $options = [])
     {
         $_documentClass = $this->_documentClass;
 
@@ -1491,7 +1590,7 @@ class CollectionHandler extends Handler
         }
 
         $body = [
-            self::OPTION_COLLECTION => $collectionId,
+            self::OPTION_COLLECTION => $this->makeCollection($collection),
             self::OPTION_EXAMPLE    => $example->getAllAsObject(['_ignoreHiddenAttributes' => true]),
             self::OPTION_NEW_VALUE  => $newValue->getAllAsObject(['_ignoreHiddenAttributes' => true])
         ];
@@ -1529,7 +1628,7 @@ class CollectionHandler extends Handler
      *
      * @throws Exception
      *
-     * @param mixed $collectionId - collection id as string or number
+     * @param mixed $collection   - collection as string or object
      * @param mixed $example      - the example document as a Document object or an array
      * @param mixed $newValue     - patch document or array which contains the attributes and values to be replaced
      * @param mixed $options      - optional, array of options (see below) or the boolean value for $policy (for compatibility prior to version 1.1 of this method)
@@ -1539,10 +1638,12 @@ class CollectionHandler extends Handler
      *                            <li>'limit'       - can be used set a limit on how many documents to replace at most. If limit is specified but is less than the number of documents in the collection, it is undefined which of the documents will be replaced.</li>
      *                            </p>
      *
+     * @deprecated use AQL queries instead
+     *
      * @return bool - always true, will throw if there is an error
      * @since 1.2
      */
-    public function replaceByExample($collectionId, $example, $newValue, array $options = [])
+    public function replaceByExample($collection, $example, $newValue, array $options = [])
     {
         $_documentClass = $this->_documentClass;
 
@@ -1555,7 +1656,7 @@ class CollectionHandler extends Handler
         }
 
         $body = [
-            self::OPTION_COLLECTION => $collectionId,
+            self::OPTION_COLLECTION => $this->makeCollection($collection),
             self::OPTION_EXAMPLE    => $example->getAllAsObject(['_ignoreHiddenAttributes' => true]),
             self::OPTION_NEW_VALUE  => $newValue->getAllAsObject(['_ignoreHiddenAttributes' => true])
         ];
@@ -1591,7 +1692,7 @@ class CollectionHandler extends Handler
      *
      * @throws Exception
      *
-     * @param mixed $collectionId      - collection id as string or number
+     * @param mixed $collection        - collection as string or object
      * @param mixed $document          - the example document as a Document object or an array
      * @param array $options           - optional - an array of options.
      *                                 <p>Options are :<br>
@@ -1606,7 +1707,7 @@ class CollectionHandler extends Handler
      *
      * @since 1.2
      */
-    public function removeByExample($collectionId, $document, array $options = [])
+    public function removeByExample($collection, $document, array $options = [])
     {
         $_documentClass = $this->_documentClass;
 
@@ -1619,7 +1720,7 @@ class CollectionHandler extends Handler
         }
 
         $body = [
-            self::OPTION_COLLECTION => $collectionId,
+            self::OPTION_COLLECTION => $this->makeCollection($collection),
             self::OPTION_EXAMPLE    => $document->getAllAsObject(['_ignoreHiddenAttributes' => true])
         ];
 
@@ -1652,7 +1753,7 @@ class CollectionHandler extends Handler
      *
      * @throws Exception
      *
-     * @param mixed $collectionId      - collection id as string or number
+     * @param mixed $collection        - collection as string or object
      * @param array $keys              - array of document keys
      * @param array $options           - optional - an array of options.
      *                                 <p>Options are :<br>
@@ -1666,10 +1767,10 @@ class CollectionHandler extends Handler
      *
      * @since 2.6
      */
-    public function removeByKeys($collectionId, array $keys, array $options = [])
+    public function removeByKeys($collection, array $keys, array $options = [])
     {
         $body = [
-            self::OPTION_COLLECTION => $collectionId,
+            self::OPTION_COLLECTION => $this->makeCollection($collection),
             self::OPTION_KEYS       => $keys
         ];
 
@@ -1705,7 +1806,7 @@ class CollectionHandler extends Handler
      *
      * @throws Exception
      *
-     * @param mixed $collectionId        - collection id as string or number
+     * @param mixed $collection          - collection as string or object
      * @param array $keys                - array of document keys
      * @param array $options             - optional array of options.
      *                                   <p>Options are :<br>
@@ -1718,12 +1819,12 @@ class CollectionHandler extends Handler
      *
      * @since 2.6
      */
-    public function lookupByKeys($collectionId, array $keys, array $options = [])
+    public function lookupByKeys($collection, array $keys, array $options = [])
     {
         $_documentClass = $this->_documentClass;
 
         $body = [
-            self::OPTION_COLLECTION => $collectionId,
+            self::OPTION_COLLECTION => $this->makeCollection($collection),
             self::OPTION_KEYS       => $keys
         ];
 
@@ -1748,7 +1849,7 @@ class CollectionHandler extends Handler
      *
      * @throws Exception
      *
-     * @param mixed  $collectionId    - collection id as string or number
+     * @param mixed  $collection      - collection as string or object
      * @param string $attribute       - the attribute path , like 'a', 'a.b', etc...
      * @param mixed  $left            - The lower bound.
      * @param mixed  $right           - The upper bound.
@@ -1769,9 +1870,11 @@ class CollectionHandler extends Handler
      *                                </li>
      *                                </p>
      *
+     * @deprecated use AQL queries instead
+     *
      * @return Cursor - documents matching the example [0...n]
      */
-    public function range($collectionId, $attribute, $left, $right, array $options = [])
+    public function range($collection, $attribute, $left, $right, array $options = [])
     {
         if ($attribute === '') {
             throw new ClientException('Invalid attribute specification');
@@ -1783,7 +1886,7 @@ class CollectionHandler extends Handler
         }
 
         $body = [
-            self::OPTION_COLLECTION => $collectionId,
+            self::OPTION_COLLECTION => $this->makeCollection($collection),
             self::OPTION_ATTRIBUTE  => $attribute,
             self::OPTION_LEFT       => $left,
             self::OPTION_RIGHT      => $right
@@ -1812,10 +1915,9 @@ class CollectionHandler extends Handler
      *
      * This will throw if the list cannot be fetched from the server
      *
-     *
      * @throws Exception
      *
-     * @param mixed  $collectionId    - collection id as string or number
+     * @param mixed  $collection      - collection as string or object
      * @param double $latitude        - The latitude of the coordinate.
      * @param double $longitude       - The longitude of the coordinate.
      * @param array  $options         - optional array of options.
@@ -1835,12 +1937,14 @@ class CollectionHandler extends Handler
      *                                </li>
      *                                </p>
      *
+     * @deprecated use AQL queries instead
+     *
      * @return Cursor - documents matching the example [0...n]
      */
-    public function near($collectionId, $latitude, $longitude, array $options = [])
+    public function near($collection, $latitude, $longitude, array $options = [])
     {
         $body = [
-            self::OPTION_COLLECTION => $collectionId,
+            self::OPTION_COLLECTION => $this->makeCollection($collection),
             self::OPTION_LATITUDE   => $latitude,
             self::OPTION_LONGITUDE  => $longitude
         ];
@@ -1871,7 +1975,7 @@ class CollectionHandler extends Handler
      *
      * @throws Exception
      *
-     * @param mixed  $collectionId    - collection id as string or number
+     * @param mixed  $collection      - collection as string or object
      * @param double $latitude        - The latitude of the coordinate.
      * @param double $longitude       - The longitude of the coordinate.
      * @param int    $radius          - The maximal radius (in meters).
@@ -1892,12 +1996,14 @@ class CollectionHandler extends Handler
      *                                </li>
      *                                </p>
      *
+     * @deprecated use AQL queries instead
+     *
      * @return Cursor - documents matching the example [0...n]
      */
-    public function within($collectionId, $latitude, $longitude, $radius, array $options = [])
+    public function within($collection, $latitude, $longitude, $radius, array $options = [])
     {
         $body = [
-            self::OPTION_COLLECTION => $collectionId,
+            self::OPTION_COLLECTION => $this->makeCollection($collection),
             self::OPTION_LATITUDE   => $latitude,
             self::OPTION_LONGITUDE  => $longitude,
             self::OPTION_RADIUS     => $radius